requirements.txt
Django
django-extensions
django-cors-headers
psycopg2-binary
Pillow
djangorestframework
django-rest-registration